Round smlv somatic FORMAT/SQ in QUAL field
The BCFtool stats section in the MultiQC report contains a frequency
distribution plot of QUAL scores, however DRAGEN does not populate this
field but instead provides FORMAT/SQ as a measure of variant quality.

In order to show FORMAT/SQ in the MultiQC report it is moved to the QUAL
field prior to running BCFtools stats. To make the frequency
distribution plot cleaner, FORMAT/SQ are now rounded before transferring
them to the QUAL field.