Switch to Bioconda base and simplify Dockerfiles
The Google Distroless images do not have numbered versioning and
typically the `latest` tag is expected to be used. However, the most
recent base-debian11 image contains a glibc version that is incompatible
with the PCGR install. Rather than trying to fix this I have switched to
the standard Bioconda base image, which also includes BusyBox binaries
allowing simplification of the Dockerfiles.